Can now extract the path from the dropbox link.
NOTE: This is probably a really bad way to do it as if dropbox change
their shared link format we’re in serious trouble.

The justification for doing it this way was that I can use the chooser,
rather than writing my own dropbox file browser.

Sleiman has informed me that my rest controller is all wrong. Fix that!