CSRF Prevention
Until now I’ve been using the no redirect version of dropbox’s
authentication. This was the one in their demo which ran in the
console. They have a different version, where you link the user to
their site and they will redirect them back to your site. This is
better because the user doesn’t have to worry about copying/pasting
authentication codes.

This method also has the benefit of being much more transparent as to
what error has been thrown in the event that it goes wrong (how I
stumbled onto it in the first place).

To implement this method you need to generate a CSRF Token (this is a
randomly generated string that is unique to a ‘session’) and give that
to the dropbox API.

When they navigate to the root directory I need to generate a new one
of these tokens. Then return that token to the client as a token and
re-direct them to dropbox (letting dropbox know the token).

Then when they return from Dropbox I need to compare the token in their
cookie to the one dropbox spits back. If they’re the same then it’s the
same person. If they’re different I don’t quite understand what an
attacker would be trying to do but apparently it’s bad.

The problem I have at the minute is that to create this DbxWebAuth
object you need to supply
- requestConfig (easy)
- appInfo (easy)
- redirector (easy)
- csrfTokenStore (not so easy)

csrfTokenStore is a DbxSessionStore object and to construct it you need
to give it a session object and a session key.

This object needs to be the same for the initial load and the redirect
from Dropbox. The way I understand it the session will be the same if
it comes from the same host so that part should be fine. This issue is
with the sessionKey. On the first call this is fine, I can generate it.
On the second call I need to somehow remember this key. Can I just
store it as a cookie and retrieve it?